Fix predictive mean matching for linear predictors.
The old implementation of predictive mean matching imputed based on the
distance between the donor values and the predicted target. However, the
method is supposed to impute based on the distance between predicted
donor values and predicted target.

This change implements the correct behaviour for predictive mean
matching using linear predictors (impute_lm, impute_rlm and impute_en).
It also stops execution with an error if any other imputation function
is passed as a predictor for impute_pmm().