Fix predictive mean matching imputation. (#42)
* Fix predictive mean matching for linear predictors.

The old implementation of predictive mean matching imputed based on the
distance between the donor values and the predicted target. However, the
method is supposed to impute based on the distance between predicted
donor values and predicted target.

This change implements the correct behaviour for predictive mean
matching using linear predictors (impute_lm, impute_rlm and impute_en).
It also stops execution with an error if any other imputation function
is passed as a predictor for impute_pmm().

* Fix issues with donor selection in impute_pmm().

There were two issues with donor selection:
- Donors with NA predictions were allowed in the donor pool, which could
  lead to break-down in extreme cases without valid donors.
- For the multivariate option, records in which values had been imputed
  in earlier iterations were excluded from the donor pool, even if the
  imputations were only in columns irrelevant to the current iteration.

This change fixes these issues and adds tests.

* Use gower::topn() instead of C routine in univariate PMM.

Multivariate PMM was already using gower::topn() to find the best donor
values. There have been no reports of performance issues, and there is
no reason to assume that performance will not suffice for the univariate
case as well. This change slightly simplifies the package.

* Add impute_all option to impute_cart() and impute_rf().

With this change, tree models can be used as predictor in impute_pmm().

* Improve impute_pmm() to handle imputed data with 0 range.

If the range is 0, gower_topn() cannot calculate the closest donor. We
catch this case and pick the first eligible donor, which is consistent
with the behaviour for gower_topn() in case of equal donor values.

* Update NEWS file.

---------

Co-authored-by: Reijer Idema <r.idema@cbs.nl>